\name{netie-package}
\alias{netie-package}
\alias{netie}
\docType{package}
\title{
\packageTitle{netie}
}
\description{
\packageDescription{netie}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{netie}
\packageIndices{netie}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{netie}

Maintainer: \packageMaintainer{netie}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
